% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_axes_labels}
\alias{add_axes_labels}
\title{Add labels to an axis at given positions}
\usage{
add_axes_labels(v, U, labels, i = 1:nrow(U), ortho = FALSE, below = TRUE,
  line = 1, srt = NULL, ...)
}
\arguments{
\item{v}{Direction vector of axis.}

\item{U}{Points on axis in user coordinates}

\item{labels}{Labels to place on axis.}

\item{i}{Indexes of which labels to plot.}

\item{ortho}{Labels orthogonal to axis? (default \code{FALSE}).}

\item{below}{Should the label appear below the axis?}

\item{line}{Number of lines label is placed below axis.}

\item{srt}{A numerical value specifying (in degrees) how the text
should be rotated.}

\item{...}{Passed on to \code{\link{text}} to plot labels.}
}
\description{
Add labels to an axis at given positions
}
\details{
Places value labels at given points and in direction given by
vector. The labels are rotated automatically to be readable.
}
\section{TODO Respect cex argument, correct char width (SO question)}{

}
\examples{
# generate data
set.seed(1)
v <- c(1,1)                                         # vector in direction of axis
X <- t(replicate(4, runif(1, -1, 1) * v))           # generate points on axis

## plot 1: add labels parallel to axis

plot(X, pch=16, asp=1, xlim=c(-1,1), ylim=c(-1,1))  
abline(h=0, v=0, col="grey")
abline(0, 1)    

# add point coords as labels
lbls <- apply(X, 1, function(x) paste(round(x,1), collapse=","))
add_axes_labels(v, X, lbls, cex=.7)


## plot 2: add labels perpendicular to axis / without rotation

plot(X, pch=16, asp=1, xlim=c(-1,1), ylim=c(-1,1))  
abline(h=0, v=0, col="grey")
abline(0, 1)  

# add projected values on vector as labels 
vn <- v / sum(v^2)^.5                               # normalize vector for projection             
lbls <- X \%*\% v                                     # project points on vector
add_axes_labels(v, X, round(lbls,1), cex=.7, 
                below=FALSE, col="blue", ortho=TRUE)
add_axes_labels(v, X, round(lbls, 1), cex=.7, srt=0, 
                line = .5, col="grey")

}

