% Generated by roxygen2 (4.0.2): do not edit by hand
\name{axes_label_offset}
\alias{axes_label_offset}
\title{Generate offset for axis labels in usr coords.}
\usage{
axes_label_offset(v, line = 1, below = TRUE)
}
\arguments{
\item{v}{Vector of axis direction.}

\item{line}{Offset in line heights}

\item{below}{Offset in direction below thes axis?}
}
\value{
A vector containing width and height for single character.
}
\description{
Generate offset for axis labels in usr coords.
}
\details{
Generates an offset to place labels below or above an axis.
The offest is given in multiple of the lineheight.
}
\examples{
axes_label_offset(c(1,1))
}

