% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biplot_visualize}
\alias{biplot_visualize}
\title{Visualization workhorse for biplots.}
\usage{
biplot_visualize(L, R)
}
\arguments{
\item{L}{Matrix with row points.}

\item{R}{Matrix with variable vectors.}
}
\value{
NULL.
}
\description{
Visualization workhorse for biplots.
}
\keyword{internal}

