% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseIP.R
\name{collapseIP}
\alias{collapseIP}
\title{Title}
\usage{
collapseIP(
  exb,
  coerce_cols = "Text",
  annotation_dummies = FALSE,
  annotation_cols = c()
)
}
\arguments{
\item{exb}{data.frame}

\item{coerce_cols}{cols that are pasted together (names)}

\item{annotation_dummies}{should the (character) annotation colums be converted to dummies and be added up per ips}

\item{annotation_cols}{character annotation colums (names)}
}
\description{
Title
}
