% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convex_hull}
\alias{convex_hull}
\title{Draw different types of convex hulls around a set of points}
\usage{
convex_hull(x, y = NULL, alpha = 0, peel = FALSE, smooth = 0,
  shape = 1, draw = TRUE, ...)
}
\arguments{
\item{x,y}{The x and y coordinates of a set of points.
Alternatively, a single argument x can be provided.}

\item{alpha}{Proportion of points most distant from centroid to
be removed before creating the (peeled) convex hull.}

\item{peel}{Remove hull points? (useful to remove outliers,
default \code{FALSE})}

\item{smooth}{Integer (1 to 3) indicating the method for creating splines
around the hull.}

\item{shape}{Shape parameter used for \code{smooth=1}
(see \code{xspline} for details).}

\item{draw}{Draw the hull polygon (default \code{TRUE}).}

\item{...}{additional parameters passed to \code{polygon} to draw the hull.}
}
\value{
A list as given by \code{\link{xy.coords}} containing the
 hull points is returned invisibly.
}
\description{
Draw different types of convex hulls around a set of points
}
\examples{

set.seed(2)
x <- rnorm(100)
y <- rnorm(100)
op <- par(mfrow=c(2,3), mar=c(3,2,3,1))

plot(x,y, main="convex hull")
convex_hull(x,y)

plot(x,y, main="colored hull")
convex_hull(x,y, col="#FF000050", density=10, angle=45, border="red")

plot(x,y, main="20\% of most distant\\npoints removed")
convex_hull(x,y, alpha=.2, col="#FF000050", density=10, angle=45, border="red")

plot(x,y, main="initial hull points\\nremoved (peeled off)")
convex_hull(x,y, peel=T, col="#00FF0050", border="red")

plot(x,y, main="smoothed hull\\n(spline interpolated)")
convex_hull(x,y, smooth=2, col="#FF000050", border="red")

plot(x,y, main="another smooting\\nmethod")
convex_hull(x,y, peel=T, smooth=1, shape=1, col="#00FF0050", border="red")

par(op)
}

