% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKWIC.R
\name{createKWIC}
\alias{createKWIC}
\title{Title}
\usage{
createKWIC(
  exb_tok = data.frame(),
  window = 2,
  pattern = "",
  textCol = "Text",
  summarise_pattern = FALSE,
  IPintern = TRUE
)
}
\arguments{
\item{exb_tok}{tokenisiertes exb objekt}

\item{window}{fenster um Zielwort/wörter herum}

\item{pattern}{Zielwort/phrase}

\item{textCol}{Zeile mit Text aufgrunddessen die KWi}

\item{summarise_pattern}{if True rows for pattern are summarised}
}
\description{
Title
}
