% Generated by roxygen2 (4.0.2): do not edit by hand
\name{degree_to_rad}
\alias{degree_to_rad}
\title{Conversion from degrees to radians}
\usage{
degree_to_rad(x)
}
\arguments{
\item{x}{Degrees.}
}
\description{
Conversion from degrees to radians
}
\examples{
degree_to_rad(seq(0, 360, length=45))
}

