% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_2d_normal_vector}
\alias{get_2d_normal_vector}
\title{Get normal vector for given 2D vector.}
\usage{
get_2d_normal_vector(v, upwards = TRUE)
}
\arguments{
\item{v}{A 2D vector.}

\item{upwards}{Vector direction is selected so it points in
positive y-direction by default.}
}
\value{
A 2D vector of length one.
}
\description{
Get normal vector for given 2D vector.
}
\examples{
v <- c(1,1)
 u <- get_2d_normal_vector(v)
 d <- get_2d_normal_vector(v, upwards=FALSE)
 plot(NULL, xlim=c(-1,1), ylim=c(-1,1), asp=1)
 arrows(0, 0, v[1], v[2])
 arrows(0, 0, u[1], u[2], col=2)
 arrows(0, 0, d[1], d[2], col=3)
}

