% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_crit_density_value}
\alias{get_crit_density_value}
\title{Calculate a critical denisty value from the 2D density distribution.
The critical value can be used.}
\usage{
get_crit_density_value(den, n = 100, prob = 0.05)
}
\arguments{
\item{den}{Object returned by \code{kde2d}.}

\item{n}{A bit uncelar needs a check.}

\item{prob}{Probability used to determine critical value.}
}
\description{
Calculate a critical denisty value from the 2D density distribution.
The critical value can be used.
}
\section{TODO}{

double check meaning of n
http://grokbase.com/t/r/r-help/1233hxv1a9/r-contour-for-plotting-confidence-interval-on-scatter-plot-of-bivariate-normal-distribution
n might have to equal nobs
http://quantitative-ecology.blogspot.de/2008/03/plotting-contours.html
}
\keyword{internal}

