% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_plot_asp_ratio}
\alias{get_plot_asp_ratio}
\title{Get plot aspect ratio and 1mm in usr coordinates}
\usage{
get_plot_asp_ratio()
}
\value{
A list with the following elements:
   \item{\code{asp}}{Plot aspect ratio}
   \item{\code{usr.1mm}}{width and height of 1 mm in usr coordinates}
}
\description{
Get plot aspect ratio and 1mm in usr coordinates
}
\examples{
plot(1:10, asp=2)
get_plot_asp_ratio()
}
\keyword{internal}

