% Generated by roxygen2 (4.0.2): do not edit by hand
\name{harmonize_text_srt}
\alias{harmonize_text_srt}
\title{Adjust text srt (rotation) to get readable directions}
\usage{
harmonize_text_srt(a, ortho = FALSE)
}
\arguments{
\item{a}{Vector of angles (anti-clockwise).}

\item{ortho}{Rotate so text is orthogonal to the axis?}
}
\description{
Adjust text srt (rotation) to get readable directions
}
\details{
srt values are mapped as follows:
0 to 90 = keep,
271 to 260 = keep,
91 to 270 = add 180 degrees.
}
\examples{
srt <- round(seq(0, 360, length=10))   # some angles

# plot text with standard rotation
plot(srt, rep(0, 10), type="n")
for (i in seq_along(srt))             
  text(srt[i], .5, label=srt[i], srt=srt[i])

# all labels in readable direction
hsrt <- harmonize_text_srt(srt)      
for (i in seq_along(srt))
  text(srt[i], 0, label=srt[i], srt=hsrt[i])

# all labels in readable direction but orthogonal to the ones before
hsrt <- harmonize_text_srt(srt, ortho = TRUE)  # harmonize rotation
for (i in seq_along(srt))
  text(srt[i],-.5, label=srt[i], srt=hsrt[i])

}

