% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAnnotations.R
\name{mergeAnnotations}
\alias{mergeAnnotations}
\title{Title}
\usage{
mergeAnnotations(
  newData,
  original,
  columnsToChange = names(newData),
  changes = c()
)
}
\arguments{
\item{newData}{newData that should be merged into old data}

\item{original}{original data}

\item{columnsToChange}{colums that are affected}

\item{changes}{vector with ids of changed rows, optional}
}
\description{
Title
}
