% Generated by roxygen2 (4.0.2): do not edit by hand
\name{norm_vec}
\alias{norm_vec}
\title{Euclidean vector norm}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{A vector or a matrix. If a matrix, each column is
treated as a vector and the Euclidean norm is calculated for each column.}
}
\description{
Euclidean vector norm
}

