% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUmap.R
\name{plotUMAP}
\alias{plotUMAP}
\title{plotUmap}
\usage{
plotUMAP(
  kwic = data.frame(a = 1:10, b = letters[1:10], c = letters[10:1]),
  vars = setdiff(names(kwic), c("IPId", "IPNumber", "EventID", "File", "Speaker",
    "TierID", "TierCategory", "Start", "End", "Start_time", "End_time", "Name", "Text",
    "pathAudio", "pathFile", "BelegID", "pos", "BelegNummer")),
  metric = "gower",
  n_neighbors = 50,
  min_dist = 0.01,
  minPts = 50,
  interactive = TRUE,
  clusterBase = c("visualization", "distMat"),
  label = "IPId"
)
}
\arguments{
\item{kwic}{daten, muuss keine kwic sein, kann auch andere daten}

\item{vars}{spaltennamen mit variablen die zur ähnlichkeitsberechnung herangezogen werden sollen}

\item{metric}{metric für die distanzmatrix, standard ist gower}

\item{n_neighbors}{Anzahl, der direkten Nachbarn eines Punkts, die in niedrieger Dimension erhalten bleiben sollen;
je größer desto besser wird globale Struktur erhalten, je kleiner desto besser lokale Struktur}

\item{min_dist}{je kleiner desto enger die cluster}

\item{minPts}{minPts pro Cluster}

\item{interactive}{soll der plot interaktiv rauskommen}

\item{clusterBase}{soll auf der Visualisierung mit Umap oder auf der Distanzmatix geclustert werden}

\item{label}{label das angezeigt wird wenn man auf einen Punkt geht}
}
\value{
plot
}
\description{
nutzt hdbscan um dichtebereiche zu finden und plottet distanzmatrix mit umap
}
