% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rad_to_degree}
\alias{rad_to_degree}
\title{Conversion from radians to degrees}
\usage{
rad_to_degree(x)
}
\arguments{
\item{x}{Radians.}
}
\description{
Conversion from radians to degrees
}
\examples{
rad_to_degree(seq(0, 2*pi, length=pi/2))
}

