% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rotation_matrix_2d}
\alias{rotation_matrix_2d}
\title{Generate a 2D rotation matrix}
\usage{
rotation_matrix_2d(alpha = 0, clockwise = TRUE)
}
\arguments{
\item{alpha}{Angle in radians, clockwise}

\item{clockwise}{Rotate clockwise?}
}
\value{
A 2D rotation matrix
}
\description{
Generate a 2D rotation matrix
}
\examples{
v <- c(0,1)                                  # point
as <- seq(0, pi/2, length=10)                # radians to rotate
plot(NULL, xlim=c(-1,1), ylim=c(-1,1), asp=1)
for (a in as)
  points(v \%*\% rotation_matrix_2d(a))
for (a in as)
  points(v \%*\% rotation_matrix_2d(a, FALSE), col="red")
}

