% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rotation_matrix_3d}
\alias{rotation_matrix_3d}
\title{Generate a all rotation planes for 3D rotation matrix}
\usage{
rotation_matrix_3d(a)
}
\arguments{
\item{a}{Angle in radians}
}
\value{
A list with rotation upon all of the three axes.
}
\description{
Generate a all rotation planes for 3D rotation matrix
}

