% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale_vec_length_to_mm}
\alias{scale_vec_length_to_mm}
\title{Scale length of vector to given mm}
\usage{
scale_vec_length_to_mm(v, mm = 10)
}
\arguments{
\item{v}{A 2D vector}

\item{mm}{Numeric. Length to standardize vector to in mm.}
}
\description{
This function will scale a vector in usr coordinates to mm
so the plotted length will be as desired. This will be particularly
useful in situtations where the plot aspect ratio is unequal to 1.
}
\examples{
par(mfrow=c(1,2))

plot(NULL, xlim=c(-1,1), ylim=c(-1,1), asp=2,
     main="vectors plot length\\n varies by direction")

rs <- seq(0, 2*pi, length=41)
v = c(1,0)
for (r in rs) {
  R = rotation_matrix_2d(r, clockwise = F)  
  n = v \%*\% R
  segments(0,0, n[1], n[2]) 
}

plot(NULL, xlim=c(-1,1), ylim=c(-1,1), asp=2,
     main="vectors have 10mm plot\\nlength in all directions")
for (r in rs) {
  R = rotation_matrix_2d(r, clockwise = F)  
  n = v \%*\% R
  b = scale_vec_length_to_mm(n, mm=10)
  segments(0,0, b[1], b[2]) 
}
}

