% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_alpha_color_value}
\alias{set_alpha_color_value}
\title{Set alpha value to color}
\usage{
set_alpha_color_value(col, alpha = 1)
}
\arguments{
\item{col}{A color in a standard format
(e.g. \code{"#000000"}, \code{"black"}, \code{1}).}

\item{alpha}{Alpha value to add to color
(default \code{=1}, i.e. opaque).}
}
\value{
hex color value.
}
\description{
Given a color in any standard format, the color is
converted to hex and an alpha value is added.
}
\examples{
set_alpha_color_value("black", .5)
 set_alpha_color_value(1:3, .5)
 set_alpha_color_value("#000000", .5)
}

