% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAnalyse.R
\name{shinyAnalyse}
\alias{shinyAnalyse}
\title{Title}
\usage{
shinyAnalyse(
  kwic = data.frame(IPId = paste("a", 1:10, sep = ""), a = 1:10, b = 10:1),
  corpus = data.frame(IPId = paste("a", 1:30, sep = ""), a = 1:30, b = 30:1),
  objectName = paste(deparse(substitute(kwic)), "_ann", sep = ""),
  pathPraat,
  pathSave = getwd(),
  seed = 7
)
}
\arguments{
\item{kwic}{concordanz}

\item{corpus}{corpus by ip}

\item{objectName}{name of object tp save to}

\item{pathPraat}{path where praat is saved}

\item{pathSave}{path to save rds}

\item{seed}{seed for random things to be repeatable}
}
\description{
Title
}
