% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTranscript.R
\name{showTranscript}
\alias{showTranscript}
\title{Title}
\usage{
showTranscript(
  data = data.frame(),
  corpus = data.frame(),
  rowNumber = 1,
  tier = c("beides", "Standard", "Transkript")
)
}
\arguments{
\item{data}{data}

\item{corpus}{corpus}

\item{rowNumber}{row from data}

\item{tier}{tier to show transkript}
}
\description{
Title
}
