% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_exb.R
\name{tokenize_exb}
\alias{tokenize_exb}
\title{Title}
\usage{
tokenize_exb(x, assign_times = TRUE, transcription_text = "Text")
}
\arguments{
\item{x}{data.frame from read_exb}

\item{assign_times}{assign time}

\item{transcription_text}{text column that should be split}
}
\value{
data.frame
}
\description{
Title
}
