% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec_angle}
\alias{vec_angle}
\title{Angle / radian between two vectors.}
\usage{
vec_angle(a, b, rad = TRUE, mode = 1)
}
\arguments{
\item{a,b}{Vector}

\item{rad}{Angular measure is radian?  (default \code{TRUE}).}

\item{mode}{1= smallest angle in [0, 180); 2=anti-clockwise angle with respect
#' vector b in [0, 360).}
}
\value{
Angle
}
\description{
Function comes in two flavors. Either it calculates the smallest angle
[0, 180) or the anti-clockwise angle with respect to the second
vector [0, 360).
}
\examples{
v <- c(1,0)                                  # point vector
as <- seq(0, 2*pi, length=10)                # radians to rotate
# mode 1: between 0 and 180 degrees, minimal angle / radian
for (a in as)         
  print(vec_angle(v \%*\% rotation_matrix_2d(a, F), v, mode=1))
# mode 2: between 0 and 360 degrees, anti-clockwise 
for (a in as)
  print(vec_angle(v \%*\% rotation_matrix_2d(a, F), v, mode=2))
}

