% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{dissertation-package}
\alias{dissertation-package}
\title{\pkg{dissertation} contains functions used in Mark Heckmann's PhD thesis.}
\description{
\pkg{dissertation} contains functions used in Mark Heckmann's PhD thesis.
}
\note{
Fuunctions used in my PhD thesis.
}
\author{
Mark Heckmann
}
\keyword{dissertation}
\keyword{package}

