% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dissertation-package.R
\docType{package}
\name{Dissertation-package}
\alias{Dissertation}
\alias{Dissertation-package}
\title{Dissertation: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
